refactor: moved pull related functions to GithubSyncManager.ts
